requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
fastapi==0.111.0
uvicorn==0.29.0
joblib==1.4.2
pyyaml==6.0.1
scipy==1.13.0
pydantic==2.7.1
pytest==8.2.0
streamlit==1.35.0